Changing components to material-ui
