+++ src/App.js
@@ -10,7 +10,6 @@ import {
 import OpenAnalysis from "./js/OpenAnalysis/OpenAnalysis";
 import CreateAnalysis from "./js/CreateAnalysis/CreateAnalysis";
 import MetaDataInput from "./js/MetaDataInput/MetaDataInput";
-import Home from "./js/Home";
 import Alert from "./js/Alerts/Alerts";
 import {MuiThemeProvider} from "@material-ui/core/styles";
 import {theme} from "./resources/config.js";
@@ -25,11 +24,10 @@ class App extends Component {
             <div className="content">
               <BrowserRouter>
                 <Switch>
-                  <Route exact path="/Home" component={Home} />
-                  <Route path="/OpenAnalysis" component={OpenAnalysis} />
+                  <Route exact path="/OpenAnalysis" component={OpenAnalysis} />
                   <Route path="/CreateAnalysis" component={CreateAnalysis} />
                   <Route path="/MetaDataInput" component={MetaDataInput} />
-                  <Redirect to="/Home" push />
+                  <Redirect to="/OpenAnalysis" push />
                 </Switch>
               </BrowserRouter>
             </div>
