--- src/js/Home.css
@@ -1,61 +1,0 @@
-.homeButton{
-  font-size: 17px;
-  float: left;
-  position:relative;
-  min-width: 150px;
-  max-width: 250px;
-  display: block;
-  margin: 1em;
-  border-radius: 15px;
-  border: none;
-  background: none;
-  color: inherit;
-  vertical-align: middle;
-  position: relative;
-  z-index: 1;
-	background: #fff;
-	color: #37474f;
-	overflow: hidden;
-	-webkit-transition: color 0.3s;
-	transition: color 0.3s;
-  outline:none;
-}
-.homeButton.button--inverted {
-	background: none;
-	color: #fff;
-}
-.homeButton::before,
-.homeButton::after {
-	content: '';
-	position: absolute;
-	height: 100%;
-	width: 100%;
-	bottom: 100%;
-	left: 0;
-	z-index: -1;
-	-webkit-transition: -webkit-transform 0.3s;
-	transition: transform 0.3s;
-	-webkit-transition-timing-function: cubic-bezier(0.75, 0, 0.125, 1);
-	transition-timing-function: cubic-bezier(0.75, 0, 0.125, 1);
-}
-.homeButton::before {
-	background: #a5d2e5;
-}
-.homeButton::after {
-	background: #3399cc;
-}
-.homeButton:hover {
-	color: #fff;
-}
-.homeButton:hover::before,
-.homeButton:hover::after {
-	-webkit-transform: translate3d(0, 100%, 0);
-	transform: translate3d(0, 100%, 0);
-}
-.homeButton:hover::after {
-	-webkit-transition-delay: 0.175s;
-	transition-delay: 0.175s;
-}
-body a:hover {
-  text-decoration: none;
-}
