+++ src/js/OpenAnalysis/OpenAnalysis.js
@@ -3,20 +3,6 @@ import isElectron from "is-electron";
 import EnhancedTable from "./Components/EnhancedTable.js";
 const ipcRenderer = window.ipcRenderer;
 
-const styles = theme => ({
-  root: {
-    width: "100%",
-    marginTop: 15
-  },
-  table: {
-    minWidth: 1020
-  },
-  tableWrapper: {
-    overflowX: "auto"
-  },
-  cellPadding: 5
-});
-
 class OpenAnalysis extends Component {
   constructor(props) {
     super(props);
@@ -75,7 +61,6 @@ class OpenAnalysis extends Component {
         <EnhancedTable
           allAnalysis={this.state.analysisData.allAnalysis}
           relationMap={this.state.analysisData.formatedRelations}
-          classes={styles}
         />
       );
     } else {
